requirements.txt
django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
python-dotenv==1.2.2
pymysql==1.1.2
pytest==9.0.3
pytest-django==4.12.0
black==26.3.1
flake8==7.3.0
